backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.8.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
